requirements.txt
# Django
Django==3.2.4
djangorestframework>=3.14.0

# Linting
flake8>=6.0.0
black==23.1.0
isort==5.12.0

# env variables
python-decouple>=3.7

# documentation
drf-yasg==1.21.5

# authentication
djangorestframework-simplejwt==5.2.2

django-cors-headers==3.13.0

python-dotenv==1.0.0

psycopg2==2.9.3

# aws ses
django-ses==3.3.0